changed scale type from float to double

+++ src/camera.h
@@ -12,12 +12,12 @@ namespace engix
 
         void update();
     public:
-        float scale() const {return _scale;}
-        void scale(float scale) {_scale = scale;}
+        double scale() const {return _scale;}
+        void scale(double scale) {_scale = scale;}
         Vector2d position() const {return _position;}
         void position(Vector2d pos) {_position = pos;}
     private:
-        float _scale = 1.0f;
+        double _scale = 1.0f;
         Vector2d _position;
     };
 }
